requirements.txt
asgiref==3.4.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0
django-cors-headers==3.10.1
django-redis==5.2.0
djangorestframework==3.12.4
gunicorn==20.1.0
packaging==21.3
psycopg2==2.9.3
pyparsing==3.0.6
python-dotenv==0.19.2
pytz==2021.3
redis==4.1.0
sentry-sdk==1.5.2
sqlparse==0.4.2
whitenoise==5.3.0
wrapt==1.13.3
